added error-handling to backend

+++ build/backend/artists.js
@@ -6,7 +6,7 @@ router.get("/", async (request, response) => {
     const artistsAsJSON = await fs.readFile("artists.json");
     const artists = JSON.parse(String(artistsAsJSON));
     console.log(artists);
-    response.send(artists);
+    response.status(200).json(artists);
 });
 router.post("/", async (request, response) => {
     //Create a artist in artists.json
@@ -15,9 +15,16 @@ router.post("/", async (request, response) => {
     newArtist.id = new Date().getTime();
     const artistsAsJSON = await fs.readFile("artists.json");
     const artists = JSON.parse(String(artistsAsJSON));
-    artists.push(newArtist);
-    fs.writeFile("artists.json", JSON.stringify(artists));
-    response.json(artists);
+    const alreadyExists = artists.find(artist => artist.name.toLowerCase() === newArtist.name.toLowerCase());
+    //Hvis artist allerede findes, så send 404. Ellers, push newArtist, opdater artists.json og send 200
+    if (alreadyExists) {
+        response.status(404).json({ error: "Artist trying to create already exists in database" });
+    }
+    else {
+        artists.push(newArtist);
+        fs.writeFile("artists.json", JSON.stringify(artists));
+        response.status(200).json(artists);
+    }
 });
 router.get("/:id", async (request, response) => {
     //Get a single artist by id from artists.json
@@ -25,7 +32,13 @@ router.get("/:id", async (request, response) => {
     const artistsAsJSON = await fs.readFile("artists.json");
     const artists = JSON.parse(String(artistsAsJSON));
     const artist = artists.find((artist) => artist.id === id);
-    response.send(artist);
+    //Hvis du kan finde en artist ud fra given id, så send 200. Ellers, send 400
+    if (artist) {
+        response.status(200).json(artist);
+    }
+    else {
+        response.status(404).json({ error: "Could not get specific artist from database" });
+    }
 });
 router.delete("/:id", async (request, response) => {
     //Delete a single artist by id from artists.json
@@ -33,9 +46,16 @@ router.delete("/:id", async (request, response) => {
     const artistsAsJSON = await fs.readFile("artists.json");
     const artists = JSON.parse(String(artistsAsJSON));
     const index = artists.findIndex((artist) => artist.id === id);
-    artists.splice(index, 1);
-    fs.writeFile("artists.json", JSON.stringify(artists));
-    response.json(artists);
+    //Hvis du ikke kan finde noget, der har en passende id, så send 404.
+    //Ellers, fjern artist på index-plads og send 200
+    if (index === -1) {
+        response.status(404).json({ error: "Could not delete specified artist from database" });
+    }
+    else {
+        artists.splice(index, 1);
+        fs.writeFile("artists.json", JSON.stringify(artists));
+        response.status(200).json(artists);
+    }
 });
 router.put("/:id", async (request, response) => {
     //Update a single artist by id from artists.json
@@ -44,6 +64,8 @@ router.put("/:id", async (request, response) => {
     const updatedArtist = request.body;
     const artistToUpdate = artists.find((artist) => artist.id === updatedArtist.id);
     console.log(updatedArtist);
+    //Hvis artistToUpdate findes, så opdater oplysninger og send 200
+    //Ellers, send 404 og giv en error
     if (artistToUpdate) {
         artistToUpdate.name = updatedArtist.name;
         artistToUpdate.gender = updatedArtist.gender;
@@ -55,8 +77,11 @@ router.put("/:id", async (request, response) => {
         artistToUpdate.image = updatedArtist.image;
         artistToUpdate.shortDescription = updatedArtist.shortDescription;
         artistToUpdate.isFavorite = updatedArtist.isFavorite;
+        fs.writeFile("artists.json", JSON.stringify(artists));
+        response.status(200).json(artists);
     }
-    fs.writeFile("artists.json", JSON.stringify(artists));
-    response.json(artists);
+    else {
+        response.status(404).json({ error: "Could not find artist from database" });
+    }
 });
 export { router };
