+++ src/backend/artists.ts
@@ -9,7 +9,7 @@ router.get("/", async (request: Request, response: Response) => {
     const artistsAsJSON = await fs.readFile("artists.json");
     const artists: Artist[] = JSON.parse(String(artistsAsJSON));
     console.log(artists);
-    response.send(artists);
+    response.status(200).json(artists);
 });
 
 router.post("/", async (request: Request, response: Response) => {
@@ -21,10 +21,16 @@ router.post("/", async (request: Request, response: Response) => {
     const artistsAsJSON = await fs.readFile("artists.json");
     const artists: Artist[] = JSON.parse(String(artistsAsJSON));
 
-    artists.push(newArtist);
+    const alreadyExists = artists.find(artist => artist.name.toLowerCase() === newArtist.name.toLowerCase())
 
-    fs.writeFile("artists.json", JSON.stringify(artists));
-    response.json(artists);
+    //Hvis artist allerede findes, så send 404. Ellers, push newArtist, opdater artists.json og send 200
+    if (alreadyExists) {
+        response.status(404).json({ error: "Artist trying to create already exists in database" });
+    } else {
+        artists.push(newArtist);
+        fs.writeFile("artists.json", JSON.stringify(artists));
+        response.status(200).json(artists);
+    }
 });
 
 router.get("/:id", async (request: Request, response: Response) => {
@@ -32,10 +38,15 @@ router.get("/:id", async (request: Request, response: Response) => {
     const id = Number(request.params.id);
     const artistsAsJSON = await fs.readFile("artists.json");
     const artists: Artist[] = JSON.parse(String(artistsAsJSON));
-
+    
     const artist = artists.find((artist) => artist.id === id);
-
-    response.send(artist);
+    
+    //Hvis du kan finde en artist ud fra given id, så send 200. Ellers, send 400
+    if (artist) {
+        response.status(200).json(artist);
+    } else {
+        response.status(404).json({ error: "Could not get specific artist from database" });
+    }
 });
 
 router.delete("/:id", async (request: Request, response: Response) => {
@@ -45,10 +56,16 @@ router.delete("/:id", async (request: Request, response: Response) => {
     const artists: Artist[] = JSON.parse(String(artistsAsJSON));
 
     const index = artists.findIndex((artist) => artist.id === id);
-    artists.splice(index, 1);
 
-    fs.writeFile("artists.json", JSON.stringify(artists));
-    response.json(artists);
+    //Hvis du ikke kan finde noget, der har en passende id, så send 404.
+    //Ellers, fjern artist på index-plads og send 200
+    if (index === -1) {
+        response.status(404).json({ error: "Could not delete specified artist from database" });
+    } else {
+        artists.splice(index, 1);
+        fs.writeFile("artists.json", JSON.stringify(artists));
+        response.status(200).json(artists);
+    }
 });
 
 router.put("/:id", async (request: Request, response: Response) => {
@@ -61,6 +78,8 @@ router.put("/:id", async (request: Request, response: Response) => {
 
     console.log(updatedArtist);
 
+    //Hvis artistToUpdate findes, så opdater oplysninger og send 200
+    //Ellers, send 404 og giv en error
     if (artistToUpdate) {
         artistToUpdate.name = updatedArtist.name;
         artistToUpdate.gender = updatedArtist.gender;
@@ -72,10 +91,13 @@ router.put("/:id", async (request: Request, response: Response) => {
         artistToUpdate.image = updatedArtist.image;
         artistToUpdate.shortDescription = updatedArtist.shortDescription;
         artistToUpdate.isFavorite = updatedArtist.isFavorite;
+
+        fs.writeFile("artists.json", JSON.stringify(artists));
+        response.status(200).json(artists);
+    } else {
+        response.status(404).json({ error: "Could not find artist from database" });
     }
 
-    fs.writeFile("artists.json", JSON.stringify(artists));
-    response.json(artists);
 });
